Feat: validate czi axes in config
