+++ src/careamics/dataset_ng/image_stack_loader/image_stack_loaders.py
@@ -13,7 +13,8 @@ from ..image_stack import (
     InMemoryImageStack,
     ZarrImageStack,
 )
-from ..image_stack.czi_image_stack import CziImageStack, are_axes_valid
+from ..image_stack.czi_image_stack import CziImageStack
+from ..image_stack.image_utils import are_czi_axes_valid
 from .zarr_utils import collect_arrays, decipher_zarr_uri, is_ome_zarr, is_valid_uri
 
 if TYPE_CHECKING:
@@ -249,7 +250,7 @@ def load_czis(
     ValueError
         If the provided axes are not valid.
     """
-    if are_axes_valid(axes) is False:
+    if are_czi_axes_valid(axes) is False:
         raise ValueError(
             f"Provided axes '{axes}' are not valid. Axes must be in the `SC(Z/T)YX` "
             f"format, where Z or T are optional, and S and C can be singleton "
