+++ tests/dataset_ng/image_stack_loader/test_czi_image_stack_loader.py
@@ -2,7 +2,7 @@ import pytest
 
 from careamics.dataset_ng.image_stack.czi_image_stack import (
     CziImageStack,
-    are_axes_valid,
+    are_czi_axes_valid,
 )
 from careamics.dataset_ng.image_stack_loader import load_czis
 
@@ -40,7 +40,7 @@ class TestCziImageStackLoader:
 
         source = ["mocked_file.czi"]
 
-        if not are_axes_valid(axes):
+        if not are_czi_axes_valid(axes):
             with pytest.raises(ValueError):
                 _ = load_czis(source=source, axes=axes)
         else:
